+++ src/components/textarea/textarea.styles.ts
@@ -43,8 +43,7 @@ export default css`
     background-color: var(--sl-input-background-color-focus);
     border-color: var(--sl-input-border-color-focus);
     color: var(--sl-input-color-focus);
-    outline: var(--sl-focus-ring-style) var(--sl-focus-ring-width) var(--sl-input-focus-ring-color);
-    outline-offset: var(--sl-input-focus-ring-offset);
+    box-shadow: 0 0 0 var(--sl-focus-ring-width) var(--sl-input-focus-ring-color);
   }
 
   .textarea--standard.textarea--focused:not(.textarea--disabled) .textarea__control {